add secrets ignore file for false positives
.bluemix/nls/messages.yml, line 18 has:
deploy.apikey: "IBM Cloud API key"
is not a secret/password value, it is just a UI string shown to the user
containing the phrase "api key".

The .bluemix/nls/messages_*.yml files are translated versions of the
messages.yml file, only translated to other languages like French,
German, etc,
so the messages_*.yml files are ok to ignore, for the same reasons that
the messages.yml file is ok to ignore.